AC-27 #comment Reduced domain module
