+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/episode/datasource/RemoteEpisodesDataSource.kt
@@ -1,8 +1,7 @@
 package com.jfmr.ac.test.data.remote.episode.datasource
 
 import com.jfmr.ac.test.data.api.rickandmorty.episode.entity.EpisodeResponse
-import com.jfmr.ac.test.domain.model.character.DomainResult
-import com.jfmr.ac.test.domain.model.episode.Episode
+import com.jfmr.ac.test.domain.model.DomainResult
 import kotlinx.coroutines.flow.Flow
 
 interface RemoteEpisodesDataSource {
