+++ domain/model/src/main/java/com/jfmr/ac/test/domain/model/DomainResult.kt
@@ -1,0 +1,6 @@
+package com.jfmr.ac.test.domain.model
+
+import arrow.core.Either
+import com.jfmr.ac.test.domain.model.error.DomainError
+
+typealias DomainResult <T> = Either<DomainError, T>
