--- domain/model/src/main/java/com/jfmr/ac/test/domain/model/character/DomainResult.kt
@@ -1,6 +1,0 @@
-package com.jfmr.ac.test.domain.model.character
-
-import arrow.core.Either
-import com.jfmr.ac.test.domain.model.error.DomainError
-
-typealias DomainResult <T> = Either<DomainError, T>
