--- domain/usecase/di/build.gradle
@@ -1,23 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-android {
-    namespace = "com.jfmr.ac.test.domain.usecase.di"
-}
-kapt {
-    correctErrorTypes true
-}
-
-dependencies {
-    api(project(':data:api',))
-    api(project(":data:di"))
-    api(project(":domain:model"))
-    implementation(project(":domain:usecase:implementation"))
-    api(project(":domain:usecase:open"))
-
-    implementation libs.bundles.hilt
-    kapt libs.bundles.hiltkapt
-
-}
