--- domain/usecase/implementation/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-
-}
-
-android {
-    namespace = "com.jfmr.ac.test.domain.usecase.implementation"
-}
-kapt {
-    correctErrorTypes true
-}
-
-dependencies {
-    api(project(":domain:model"))
-    api api(project(':data:api',))
-    api(project(":data:di"))
-    api(project(":domain:repository"))
-    api(project(":domain:usecase:open"))
-    api(project(path: ":data:cache"))
-
-    implementation libs.bundles.hilt
-    kapt libs.bundles.hiltkapt
-
-    implementation libs.bundles.arrow
-    kapt libs.arrow.kapt
-}
