--- domain/usecase/open/.gitignore
@@ -1,1 +1,0 @@
-/build
