--- domain/usecase/open/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-android {
-    namespace = "com.jfmr.ac.test.domain.usecase.open"
-}
-
-dependencies {
-    api(project(path: ":domain:model"))
-    api(project(path: ':data:api'))
-
-    implementation libs.bundles.arrow
-    kapt libs.arrow.kapt
-
-    implementation libs.bundles.hilt
-    kapt libs.bundles.hiltkapt
-}
