--- domain/usecase/open/src/main/java/com/jfmr/ac/test/domain/usecase/open/character/CharacterDetailUseCase.kt
@@ -1,12 +1,0 @@
-package com.jfmr.ac.test.domain.usecase.open.character
-
-import com.jfmr.ac.test.domain.model.character.Character
-import com.jfmr.ac.test.domain.model.error.DomainError
-
-interface CharacterDetailUseCase {
-    suspend fun invoke(
-        characterId: Int,
-        success: (Character) -> Unit,
-        error: (DomainError) -> Unit,
-    )
-}
