--- domain/usecase/open/src/main/java/com/jfmr/ac/test/domain/usecase/open/character/CharactersUseCase.kt
@@ -1,9 +1,0 @@
-package com.jfmr.ac.test.domain.usecase.open.character
-
-import androidx.paging.PagingData
-import com.jfmr.ac.test.domain.model.character.Character
-import kotlinx.coroutines.flow.Flow
-
-interface CharactersUseCase {
-    fun characters(): Flow<PagingData<Character>>
-}
