--- domain/usecase/open/src/main/java/com/jfmr/ac/test/domain/usecase/open/character/UpdateCharacterUseCase.kt
@@ -1,8 +1,0 @@
-package com.jfmr.ac.test.domain.usecase.open.character
-
-import com.jfmr.ac.test.domain.model.character.Character
-
-interface UpdateCharacterUseCase {
-
-    suspend fun updateCharacter(character: Character): Character
-}
