--- domain/usecase/open/src/main/java/com/jfmr/ac/test/domain/usecase/open/episode/GetEpisodesUseCase.kt
@@ -1,12 +1,0 @@
-package com.jfmr.ac.test.domain.usecase.open.episode
-
-import com.jfmr.ac.test.domain.model.episode.Episode
-import com.jfmr.ac.test.domain.model.error.DomainError
-
-interface GetEpisodesUseCase {
-    suspend operator fun invoke(
-        episodesList: List<String>,
-        success: (List<Episode>?) -> Unit,
-        error: (DomainError) -> Unit,
-    )
-}
