+++ domain/usecase/src/main/java/com/jfmr/ac/test/usecase/character/UpdateCharacterUseCase.kt
@@ -1,0 +1,8 @@
+package com.jfmr.ac.test.usecase.character
+
+import com.jfmr.ac.test.domain.model.character.Character
+
+interface UpdateCharacterUseCase {
+
+    suspend fun updateCharacter(character: Character): Character
+}
