+++ presentation/ui/src/main/java/com/jfmr/ac/test/presentation/ui/character/detail/viewmodel/DetailViewModel.kt
@@ -7,15 +7,15 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jfmr.ac.test.domain.model.character.Character
-import com.jfmr.ac.test.domain.usecase.di.CharacterDetail
-import com.jfmr.ac.test.domain.usecase.di.UpdateCharacter
-import com.jfmr.ac.test.domain.usecase.open.character.CharacterDetailUseCase
-import com.jfmr.ac.test.domain.usecase.open.character.UpdateCharacterUseCase
 import com.jfmr.ac.test.presentation.ui.R
 import com.jfmr.ac.test.presentation.ui.character.detail.model.CharacterDetailError
 import com.jfmr.ac.test.presentation.ui.character.detail.model.CharacterDetailEvent
 import com.jfmr.ac.test.presentation.ui.character.detail.model.CharacterDetailState
 import com.jfmr.ac.test.presentation.ui.character.detail.model.mapper.CharacterDetailMapper
+import com.jfmr.ac.test.usecase.character.CharacterDetailUseCase
+import com.jfmr.ac.test.usecase.character.UpdateCharacterUseCase
+import com.jfmr.ac.test.usecase.di.CharacterDetail
+import com.jfmr.ac.test.usecase.di.UpdateCharacter
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
