+++ presentation/ui/src/main/java/com/jfmr/ac/test/presentation/ui/character/list/viewmodel/CharacterListViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.jfmr.ac.test.domain.model.character.Character
-import com.jfmr.ac.test.domain.usecase.di.GetCharacters
-import com.jfmr.ac.test.domain.usecase.di.UpdateCharacter
-import com.jfmr.ac.test.domain.usecase.open.character.CharactersUseCase
-import com.jfmr.ac.test.domain.usecase.open.character.UpdateCharacterUseCase
+import com.jfmr.ac.test.usecase.character.CharactersUseCase
+import com.jfmr.ac.test.usecase.character.UpdateCharacterUseCase
+import com.jfmr.ac.test.usecase.di.GetCharacters
+import com.jfmr.ac.test.usecase.di.UpdateCharacter
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
